+++ packages/aws-amplify-react-native/src/AmplifyTheme.ts
@@ -13,6 +13,9 @@
 
 import { StyleSheet } from 'react-native';
 
+// TODO: Add more specific theme object with keys
+export type AmplifyThemeType = Record<string, any>;
+
 // Colors
 export const deepSquidInk = '#152939';
 export const linkUnderlayColor = '#FFF';
@@ -20,6 +23,8 @@ export const errorIconColor = '#DD3F5B';
 export const textInputColor = '#000000';
 export const textInputBorderColor = '#C4C4C4';
 export const placeholderColor = '#C7C7CD';
+export const buttonColor = '#ff9900';
+export const disabledButtonColor = '#ff990080';
 
 // Theme
 export default StyleSheet.create({
@@ -35,11 +40,18 @@ export default StyleSheet.create({
 	section: {
 		flex: 1,
 		width: '100%',
-		padding: 20,
+		justifyContent: 'space-between',
+		paddingHorizontal: 20,
+	},
+	sectionScroll: {
+		flex: 1,
+		width: '100%',
+		paddingHorizontal: 20,
 	},
 	sectionHeader: {
 		width: '100%',
 		marginBottom: 32,
+		paddingTop: 20,
 	},
 	sectionHeaderText: {
 		color: deepSquidInk,
@@ -56,7 +68,13 @@ export default StyleSheet.create({
 	},
 	sectionFooterLink: {
 		fontSize: 14,
-		color: '#ff9900',
+		color: buttonColor,
+		alignItems: 'baseline',
+		textAlign: 'center',
+	},
+	sectionFooterLinkDisabled: {
+		fontSize: 14,
+		color: disabledButtonColor,
 		alignItems: 'baseline',
 		textAlign: 'center',
 	},
@@ -89,12 +107,12 @@ export default StyleSheet.create({
 		width: '100%',
 	},
 	button: {
-		backgroundColor: '#ff9900',
+		backgroundColor: buttonColor,
 		alignItems: 'center',
 		padding: 16,
 	},
 	buttonDisabled: {
-		backgroundColor: '#ff990080',
+		backgroundColor: disabledButtonColor,
 		alignItems: 'center',
 		padding: 16,
 	},
@@ -135,5 +153,9 @@ export default StyleSheet.create({
 	},
 	pickerItem: {
 		height: 44,
+	},
+	signedOutMessage: {
+		textAlign: 'center',
+		padding: 20,
 	},
 });