+++ packages/aws-amplify-react-native/src/Auth/ConfirmSignUp.tsx
@@ -20,15 +20,25 @@ import {
 	Header,
 	ErrorRow,
 	AmplifyButton,
+	SignedOutMessage,
 	Wrapper,
 } from '../AmplifyUI';
-import AuthPiece from './AuthPiece';
+import AuthPiece, { IAuthPieceProps, IAuthPieceState } from './AuthPiece';
 import TEST_ID from '../AmplifyTestIDs';
 
 const logger = new Logger('ConfirmSignUp');
 
-export default class ConfirmSignUp extends AuthPiece {
-	constructor(props) {
+interface IConfirmSignUpProps extends IAuthPieceProps {}
+
+interface IConfirmSignUpState extends IAuthPieceState {
+	code: string | null;
+}
+
+export default class ConfirmSignUp extends AuthPiece<
+	IConfirmSignUpProps,
+	IConfirmSignUpState
+> {
+	constructor(props: IConfirmSignUpProps) {
 		super(props);
 
 		this._validAuthStates = ['confirmSignUp'];
@@ -74,45 +84,48 @@ export default class ConfirmSignUp extends AuthPiece {
 		return (
 			<Wrapper>
 				<View style={theme.section}>
-					<Header theme={theme} testID={TEST_ID.AUTH.CONFIRM_SIGN_UP_TEXT}>
-						{I18n.get('Confirm Sign Up')}
-					</Header>
-					<View style={theme.sectionBody}>
-						{this.renderUsernameField(theme)}
-						<FormField
-							theme={theme}
-							onChangeText={text => this.setState({ code: text })}
-							label={I18n.get('Confirmation Code')}
-							placeholder={I18n.get('Enter your confirmation code')}
-							required={true}
-							testID={TEST_ID.AUTH.CONFIRMATION_CODE_INPUT}
-						/>
-						<AmplifyButton
-							theme={theme}
-							text={I18n.get('Confirm')}
-							onPress={this.confirm}
-							disabled={!username || !this.state.code}
-							testID={TEST_ID.AUTH.CONFIRM_BUTTON}
-						/>
+					<View>
+						<Header theme={theme} testID={TEST_ID.AUTH.CONFIRM_SIGN_UP_TEXT}>
+							{I18n.get('Confirm Sign Up')}
+						</Header>
+						<View style={theme.sectionBody}>
+							{this.renderUsernameField(theme)}
+							<FormField
+								theme={theme}
+								onChangeText={text => this.setState({ code: text })}
+								label={I18n.get('Confirmation Code')}
+								placeholder={I18n.get('Enter your confirmation code')}
+								required={true}
+								testID={TEST_ID.AUTH.CONFIRMATION_CODE_INPUT}
+							/>
+							<AmplifyButton
+								theme={theme}
+								text={I18n.get('Confirm')}
+								onPress={this.confirm}
+								disabled={!username || !this.state.code}
+								testID={TEST_ID.AUTH.CONFIRM_BUTTON}
+							/>
+						</View>
+						<View style={theme.sectionFooter}>
+							<LinkCell
+								theme={theme}
+								onPress={this.resend}
+								disabled={!this.state.username}
+								testID={TEST_ID.AUTH.RESEND_CODE_BUTTON}
+							>
+								{I18n.get('Resend code')}
+							</LinkCell>
+							<LinkCell
+								theme={theme}
+								onPress={() => this.changeState('signIn')}
+								testID={TEST_ID.AUTH.BACK_TO_SIGN_IN_BUTTON}
+							>
+								{I18n.get('Back to Sign In')}
+							</LinkCell>
+						</View>
+						<ErrorRow theme={theme}>{this.state.error}</ErrorRow>
 					</View>
-					<View style={theme.sectionFooter}>
-						<LinkCell
-							theme={theme}
-							onPress={this.resend}
-							disabled={!this.state.username}
-							testID={TEST_ID.AUTH.RESEND_CODE_BUTTON}
-						>
-							{I18n.get('Resend code')}
-						</LinkCell>
-						<LinkCell
-							theme={theme}
-							onPress={() => this.changeState('signIn')}
-							testID={TEST_ID.AUTH.BACK_TO_SIGN_IN_BUTTON}
-						>
-							{I18n.get('Back to Sign In')}
-						</LinkCell>
-					</View>
-					<ErrorRow theme={theme}>{this.state.error}</ErrorRow>
+					<SignedOutMessage {...this.props} />
 				</View>
 			</Wrapper>
 		);